Move Context fully into OpenRCT2 namespace; types into Input.h
