+++ src/openrct2-ui/interface/Widget.cpp
@@ -13,7 +13,6 @@
 #include <cmath>
 #include <openrct2-ui/UiStringIds.h>
 #include <openrct2-ui/windows/Windows.h>
-#include <openrct2/Context.h>
 #include <openrct2/Diagnostic.h>
 #include <openrct2/Game.h>
 #include <openrct2/Input.h>
