+++ src/openrct2/interface/Chat.cpp
@@ -10,6 +10,7 @@
 #include "Chat.h"
 
 #include "../Context.h"
+#include "../Input.h"
 #include "../audio/Audio.h"
 #include "../audio/AudioMixer.h"
 #include "../core/UTF8.h"
